+++ gen/paths.ts
@@ -1,7 +1,7 @@
 import fs from 'node:fs/promises';
-import path from 'node:path';
+import readline from 'node:readline';
 
-const pathsFilePath = path.join('./paths.ts');
+const pathsFilePath = './paths.ts';
 const pathsTemplate = `import { ContentType } from './types.ts';
 
 export const TEMPLATES_PATH = '';
@@ -11,10 +11,36 @@ export const CONTENT_PATHS: Record<ContentType, string> = {};`;
 await generatePaths();
 
 /**
- * @description Generate a new `paths.ts` file in the project's root.
+ * @description Generates a new `paths.ts` file in the project's root.
  *
- * This file can be committed into Git, since its values are vault-specific.
+ * The resulting file can be committed into Git, since its values are vault-specific.
  */
 async function generatePaths() {
-  await fs.writeFile(pathsFilePath, pathsTemplate);
+  try {
+    await fs.access(pathsFilePath);
+
+    const rl = readline.createInterface({
+      input: process.stdin,
+      output: process.stdout,
+    });
+
+    rl.question(
+      `A paths file already exists at "${pathsFilePath}". Would you like to override it? [y/n] `,
+      async (answer) => {
+        if (answer === 'y') {
+          console.log(`Overwriting the paths file at "${pathsFilePath}".`);
+          await fs.writeFile(pathsFilePath, pathsTemplate);
+          rl.close();
+          process.exit();
+        } else {
+          rl.close();
+          process.exit();
+        }
+      }
+    );
+  } catch {
+    console.log(`Generating a new paths file at "${pathsFilePath}".`);
+    await fs.writeFile(pathsFilePath, pathsTemplate);
+    process.exit();
+  }
 }
